Additional loader options
Adds the `psc-make` options `comments` and `no-prefix`. Also, an
internal option `src` has been added that is used to specify the source
paths of `PureScript` files that will be globbed for compilation.

By default the `bower_components` path is globbed, but the loader
requires that the source paths be provided for the user's code.

Resolves #12