Always resolve on psc-ide rebuild failure
The objective of enabling psc-ide is to obtain feedback immediately from
the PureScript compiler. When the loader triggers an error to webpack,
the bundle is invalidated and dependencies of the PureScript file that
caused the error will be passed through the loader on the next
successful rebuild.

However, this behaviour adds time to the feedback loop. In order to
reduce the time of the PureScript compiler feedback loop when in
development, this commit always resolves the rebuild process with a
success. Errors and warnings are still emitted, but the webpack bundle
is not invalidated. This means that the loader will only be run for the
current file being editted and the developer gets immediate feedback on
success or failure when psc-ide is enabled.

Resolves #76